server/requirements.txt
# Flask core dependencies
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Cors==5.0.0  # Updated to the latest version

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7  # For PostgreSQL, if you ever want to switch
# For SQLite (built-in support, no external package required)

# Cloudinary integration
cloudinary==1.31.0

# Python-dotenv for environment variable management
python-dotenv==1.0.0

# Alembic for database migrations (required by Flask-Migrate)
Alembic==1.10.0

# For testing and development
pytest==7.4.0
pytest-flask==1.2.0

# Optional for improved security and password hashing
bcrypt==4.0.1

# Other useful tools
gunicorn==20.1.0  # If deploying with Gunicorn server
Werkzeug==2.3.3